REFACTOR: move explainer classes to dedicated 'explainer' sub-package
