+++ src/facet/explainer/base/__init__.py
@@ -1,0 +1,5 @@
+"""
+Abstract base classes for the `facet.explainer` package.
+"""
+
+from ._base import *
