+++ src/facet/explainer/parallel/__init__.py
@@ -1,0 +1,5 @@
+"""
+Running multiple explainers in parallel using a :class:`ParallelExplainer` instance.
+"""
+
+from ._parallel import *
